requirements.txt
Django==3.2.4
asgiref==3.3.4
attrs==21.2.0
chardet==4.0.0
charset-normalizer==2.0.3
certifi==2021.5.30
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-easy-select2==1.5.7
django-excel==0.0.10
django-import-export==2.6.0
django-reset-migrations==0.4.0
djangorestframework==3.12.4
dj-database-url==0.5.0
django-reset-migrations==0.4.0
django-cors-headers==3.8.0
django-tables2==2.4.0
et-xmlfile==1.1.0
gunicorn==20.1.0
httpie==2.4.0
idna==3.2
jsonschema==3.2.0
lml==0.1.0
MarkupPy==1.14
mysqlclient==2.0.3
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.9.1
pandas==1.2.4
pyrsistent==0.17.3
pytz==2021.1
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyrsistent==0.17.3
PySocks==1.7.1
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
scipy==1.6.3
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
texttable==1.6.4
urllib3==1.26.6
whitenoise==5.3.0
xlrd==1.2.0
xlwt==1.3.0










